RC_Channel: use float for stick mixing

+++ libraries/RC_Channel/RC_Channel.h
@@ -71,7 +71,7 @@ public:
     void       set_override(const uint16_t v, const uint32_t timestamp_ms);
     bool       has_override() const;
 
-    int16_t    stick_mixing(const int16_t servo_in);
+    float    stick_mixing(const float servo_in);
 
     // get control input with zero deadzone
     int16_t    get_control_in_zero_dz(void) const;
